pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tender_normalizer_py"
version = "0.1.0"
description = "Python-based tender normalization system using Pydantic and PydanticAI"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-ai>=0.0.31",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",  # Keep for local development, will be ignored on Apify
    "openai>=1.10.0",
    "logfire>=0.20.0",  # For monitoring
    "asyncpg>=0.29.0",  # For Supabase connection
    "langdetect>=1.0.9",  # For language detection
    "httpx>=0.25.0",  # Used by logfire instrumentation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.5",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py310" 